pyproject.toml
[project]
name = "example-agents"
version = "0.1.0"
description = "Example Agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.23.0",
    "google-cloud-aiplatform>=1.75.0",
    "google-cloud-logging>=3.13.0",
    "google-genai>=0.3.0",
    "langchain-community==0.3.20",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-gcp-trace>=1.11.0",
    "opentelemetry-sdk>=1.37.0",
    "python-dotenv>=1.2.1",
    "wikipedia==1.4.0",
    "yfinance>=1.1.0",
]
[tool.ruff.lint]
select = ["E", "F", "D", "I"]
ignore = ["D203", "D212", "E501", "F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"
